Css refactoring final

+++ app/Http/Controllers/ReviewController.php
@@ -2,9 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use App\Models\Review;
 use Illuminate\Support\Facades\Auth;
+
+// The ReviewController contains the functions related to the reviews system
 
 class ReviewController extends Controller
 {
@@ -30,7 +33,8 @@ class ReviewController extends Controller
         // Validate the rating and comment fields
         $request->validate([
             'rating' => 'required|integer|between:1,5',
-            'contents' => 'nullable|string',
+            'contents' => 'nullable|string|max:255',
+            'title' => 'required|string|max:255'
         ]);
 
         // Create the review and associate it with the authenticated user and the product being reviewed
@@ -45,17 +49,16 @@ class ReviewController extends Controller
         return redirect()->back()->with('success', 'Review submitted successfully!');
     }
 
-    public function destroy($id)
+    public function deleteReview($id): RedirectResponse
     {
-        $review = Review::where('id', '=', $id);
-        // Check if the authenticated user is authorized to delete the review
-        // if ($review->user_id === auth()->id()) {
+        $review = Review::findOrFail($id);
+        if (( $review->user_id == Auth::id() ) || ( Auth::user()->accountType == 1 )){
             $review->delete();
-            return redirect(url('/products/'.$id))->with('success', 'Review deleted successfully.');
-        // } else {
-        //     return redirect()->back()->with('error', 'You are not authorized to delete this review.');
-        // }
+            return redirect()->back()->with('success', 'review-deleted');
+        } else {
+            return redirect()->back()->with('failed', 'unauthorised');
+        }
 
-        
+
     }
 }
